Added Hann window taper in order to avoid ringing in situations where the signal is terminated as a result of the frequency decreasing with increasing time (when this happens, the validity of the Post-Newtonian expansion questionable). This taper uses half of a Hann window which zero at some time t0 and unity for some other time t1. Values outside that range are set to either zero or one, depending on whether t0 or t1 is closer (so that the window is continuous and smooth at t0 and t1). It is evaluated using the function halfhann(t,t0,t1), which has been added to Tools.c, with a prototype added to Declarations.h.
In this case, the taper is evaluated with t0 set to tend, the latest time for
which hp and hc are set to non-zero values (Or, technically, the first time for
which they are set to zero due to the Post-Newtonian expansion breaking down or
the end of the observation interval being reached). t1 is set to t0-3.0/fold,
where fold is the highest frequency reached before the orbit integration
terminated. In this scheme, the taper at times prior to tend depends on the
value of tend. Rather than attempting to determine the value of tend before the
loop evaluating hp and hc has actually reached it, that loop now begins with the
final time and works backwards, updating tend as it goes.

The deficiency in the original taper is not evident for the binary parameters in
test.c, but can be seen if the definition of the binary parameters is replaced
with the following:

SBH.Mass1 = 2.0e6;
SBH.Mass2 = 1.2e6;
SBH.CoalescenceTime = 31458149.7632;
SBH.Distance = 8.415e10;
SBH.Spin1 = 0.7;
SBH.Spin2 = 0.4;
SBH.EclipticLatitude = acos(-7.4092914023e-02);
SBH.InitialPolarAngleL = acos(8.7804266048e-01);
SBH.PolarAngleOfSpin1 = acos(-7.4556848983e-01);
SBH.PolarAngleOfSpin2 = acos(-1.6813779310e-01);
SBH.EclipticLongitude = 3.3535808424e+00;
SBH.InitialAzimuthalAngleL = 6.7510361739e-01;
SBH.AzimuthalAngleOfSpin1 = 2.5112072969e+00;
SBH.AzimuthalAngleOfSpin2 = 4.0864015221e+00;
SBH.PhaseAtCoalescence = 1.7009716373e-01;
SBH.TaperApplied = 7.0;
SBH.AmplPNorder = 0.0;
SBH.TimeSample = 15.0;
SBH.Tobs = 3.15581497632e7;
SBH.Tpad = 900.0;
SBH.Rmin = 6.0;
SBH.TaperSteepness = 150.0;



git-svn-id: https://lisatools.googlecode.com/svn@1030 c0d0b04f-df38-0410-a060-df9e05943d0e